requirements.txt
django
djangorestframework
psycopg2
markdown
django-filter
djangorestframework-simplejwt